requirements.txt
numpy==1.21.4
Pillow==9.0.0
streamlit==1.3.1
tensorflow==2.7.0
tensorflow_gpu==2.7.0
